requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
python-decouple
Pillow
django-filter